Circle tweak
Realized the usual fade wasn't present on circles. Inspection revealed this as the cause

+++ src/main/shapes/Circle.java
@@ -26,7 +26,7 @@ public class Circle extends Shape {
 
 	@Override
 	public Shape scale(double scaleModifier) {
-		return new Circle(new Point2D.Double(center.x, center.y), (int)(radius*scaleModifier));
+		return new Circle(new Point2D.Double(center.x, center.y), radius*scaleModifier);
 	}
 
 	@Override
